Asset size and usableSize are changed to BigDecimal

+++ microservices/asset/src/main/java/com/inghubs/asset/dto/AssetDto.java
@@ -1,9 +1,11 @@
 package com.inghubs.asset.dto;
 
+import java.math.BigDecimal;
+
 public record AssetDto(
         Long id,
         Long customerId,
         String assetName,
-        int size,
-        int usableSize
+        BigDecimal size,
+        BigDecimal usableSize
 ) {}
